% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_table.R
\name{closure_table}
\alias{closure_table}
\title{Make a closure table from an adjacency list}
\usage{
closure_table(adj_list, parent_col = "parent", child_col = "child")
}
\description{
Make a closure table from an adjacency list
}
\examples{
alist <- tibble::tibble(parent = c(1, 2, 2), child = c(2, 3, 4))
closure_table(alist)
}
