% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root-leaf.R
\name{filter_root_leaf}
\alias{filter_root_leaf}
\title{Filter a closure table for the root-to-leaf connections only}
\usage{
filter_root_leaf(clsr_tbl)
}
\description{
Filter a closure table for the root-to-leaf connections only
}
\examples{
alist <- tibble::tibble(parent = c(1, 2, 2), child = c(2, 3, 4))
clsr <- closure_table(alist)
filter_root_leaf(clsr)
}
