% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ggraph_nestedcats_tree}
\alias{ggraph_nestedcats_tree}
\title{Default dendrogram for visualizing nested categories}
\usage{
ggraph_nestedcats_tree(g)
}
\description{
Default dendrogram for visualizing nested categories
}
