% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-conversions.R
\name{graph_convert_edgelistdf_to_parentdf}
\alias{graph_convert_edgelistdf_to_parentdf}
\title{Convert a dataframe of edges (from, to) into a parent df (parent, node)}
\usage{
graph_convert_edgelistdf_to_parentdf(edges_df)
}
\arguments{
\item{edges_df}{a dataframe with 'from' and 'to' columns with node names as values. To be convertible, each node should have a single parent.}
}
\description{
Convert a dataframe of edges (from, to) into a parent df (parent, node)
}
\examples{
ghg_edges_df <- read_csv("data-raw/ghg_cats_edges.csv")
graph_convert_edgelistdf_to_parentdf(ghg_edges_df)
}
