% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_cats.R
\name{in_cats}
\alias{in_cats}
\title{Check if categories are inside *which* other categories}
\usage{
in_cats(v, cats, g)
}
\description{
Check if categories are inside *which* other categories
}
\examples{
in_cats(c("CH4", "NF3"), c("HFCs", "F-GHGs"), ghg_cats)
}
