% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcats-package.R
\docType{package}
\name{nestedcats-package}
\alias{nestedcats}
\alias{nestedcats-package}
\title{nestedcats: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
