% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-order-factor.R
\name{pofct}
\alias{pofct}
\title{Helper for more convenient partial order factor creation}
\usage{
pofct(x = character(), levels, partial_order = tibble(), ...)
}
\arguments{
\item{x}{the values}

\item{levels}{optional character vector, mostly for ordering.}

\item{partial_order}{tibble with symbolic columns `from` and `to`}
}
\description{
Helper for more convenient partial order factor creation
}
