% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_cats_collapsible_tree}
\alias{viz_cats_collapsible_tree}
\title{Visualize nested categories as a collapsible tree}
\usage{
viz_cats_collapsible_tree(g)
}
\value{
html widget, I think
}
\description{
Visualize nested categories as a collapsible tree
}
