% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_cats_ggraph_tree}
\alias{viz_cats_ggraph_tree}
\title{Default dendrogram for visualizing nested categories}
\usage{
viz_cats_ggraph_tree(g)
}
\arguments{
\item{g}{graph of nested categories (igraph, tidygraph, others?)}
}
\description{
Default dendrogram for visualizing nested categories
}
