% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-order-factor.R
\name{new_pofct}
\alias{new_pofct}
\title{Low-level constructor for Partially Ordered Factors `pofct` S3 class}
\usage{
new_pofct(x = integer(), levels = character(), partial_order = tibble())
}
\arguments{
\item{x}{integer vector to index into levels and partial order}

\item{levels}{character vector of labels}

\item{partial_order}{an edges data frame representing the partial order.

The partial order is a data frame of edges (e.g., with integer columns `from` and `to`).
Each edge represents a has-subset relation, e.g. `to` <= `from`, which is sort of a fat arrow.}
}
\description{
Low-level constructor for Partially Ordered Factors `pofct` S3 class
}
\examples{
new_pofct(c(1L, 2L, 3L), levels = c("whole", "part1", "part2"),
  partial_order = tibble(from = c(1L, 1L), to = c(2L, 3L)))
}
