% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-order-factor.R
\name{new_pofct}
\alias{new_pofct}
\title{Low-level constructor for Partially Ordered Factors `pofct` S3 class}
\usage{
new_pofct(x = character(), po = tibble(from = character(), to = character()))
}
\arguments{
\item{x}{integer vector to index into levels and partial order}

\item{po}{an edges data frame representing the partial order.

The partial order is a data frame of edges (e.g., with integer columns `from` and `to`).
Each edge represents a has-subset relation, e.g. `to` <= `from`, which is sort of a fat arrow.}
}
\description{
Low-level constructor for Partially Ordered Factors `pofct` S3 class
}
\examples{
new_pofct(c("whole", "part1", "part2"), po = tibble(from = c("whole", "whole"), to = c("part1", "part2")))
}
