% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-order-factor.R
\name{pofct}
\alias{pofct}
\title{Helper for more convenient partial order factor creation}
\usage{
pofct(x = character(), po = tibble(from = character(), to = character()))
}
\arguments{
\item{x}{the values}

\item{levels}{optional character vector, mostly for ordering.}

\item{partial_order}{tibble with symbolic columns `from` and `to`}
}
\description{
Helper for more convenient partial order factor creation
}
\examples{
pofct(c("fruit", "apples", "bananas"), po = tibble(from=c("fruit", "fruit"),to=c("apples", "bananas")))
}
